feat: use float for memory manegement

+++ src/cpp/hierarchical_umap.h
@@ -64,21 +64,21 @@ namespace py = pybind11;
 namespace humap {
 
 // converts py array to dense representation
-vector<vector<double>> convert_to_vector(const py::array_t<double>& v);
+vector<vector<float>> convert_to_vector(const py::array_t<float>& v);
 
 // creates a sparse object from rows, columns, and values
-vector<utils::SparseData> create_sparse(int n, const vector<int>& rows, const vector<int>& cols, const vector<double>& vals);
+vector<utils::SparseData> create_sparse(int n, const vector<int>& rows, const vector<int>& cols, const vector<float>& vals);
 
 // returns how many times each data point was an endpoint after a markov chain
-vector<int>  markov_chain(vector<vector<int>>& knn_indices, vector<double>& vals, vector<int>& cols, int num_walks, int walk_length, bool reproducible); 
+vector<int>  markov_chain(vector<vector<int>>& knn_indices, vector<float>& vals, vector<int>& cols, int num_walks, int walk_length, bool reproducible); 
 
 // returns the endpoint after a random walk
-int random_walk(int vertex, int n_neighbors, vector<double>& vals, vector<int>& cols, int walk_length, 
-	            uniform_real_distribution<double>& unif, std::mt19937& rng);
+int random_walk(int vertex, int n_neighbors, vector<float>& vals, vector<int>& cols, int walk_length, 
+	            uniform_real_distribution<float>& unif, std::mt19937& rng);
 
 
 // returns the max neighborhood after markov chain
-int markov_chain(vector<vector<int>>& knn_indices, vector<double>& vals, vector<int>& cols, 
+int markov_chain(vector<vector<int>>& knn_indices, vector<float>& vals, vector<int>& cols, 
 	             int num_walks, int walk_length, vector<int>& landmarks, int influence_neighborhood, 
 				 vector<vector<int>>& neighborhood, 
 				//  vector<vector<int>>& association, 
@@ -86,8 +86,8 @@ int markov_chain(vector<vector<int>>& knn_indices, vector<double>& vals, vector<
 				 bool reproducible);
 
 // returns the endpoint after a random walk
-int random_walk(int vertex, int n_neighbors, vector<double>& vals, vector<int>& cols, 
-				int walk_length, uniform_real_distribution<double>& unif, mt19937& rng, vector<int>& is_landmark);	
+int random_walk(int vertex, int n_neighbors, vector<float>& vals, vector<int>& cols, 
+				int walk_length, uniform_real_distribution<float>& unif, mt19937& rng, vector<int>& is_landmark);	
 
 
 /**
@@ -105,7 +105,7 @@ struct Metadata {
 	* @param association_ Container representing which data points in the level below are associated with it (TODO: use the inverse)
 	* @param size_ int representing the number of data points in the level
 	*/	
-	Metadata(vector<int> indices_, vector<int> owners_, vector<double> strength_, vector<vector<int>> association_, int size_)
+	Metadata(vector<int> indices_, vector<int> owners_, vector<float> strength_, vector<vector<int>> association_, int size_)
 	: indices(indices_), owners(owners_), strength(strength_), association(association_), size(size_)
 	{	 	
 	}
@@ -113,7 +113,7 @@ struct Metadata {
 	Metadata(vector<int> indices_, int size_): indices(indices_), size(size_) 
 	{
 		owners = vector<int>(size_, -1);
-		strength = vector<double>(size_, -1.0);
+		strength = vector<float>(size_, -1.0);
 		association = vector<vector<int>>(size_, vector<int>());
 	}
 
@@ -124,7 +124,7 @@ struct Metadata {
 	vector<int> owners;
 	vector<int> count_influence;
 
-	vector<double> strength;
+	vector<float> strength;
 
 	vector<vector<int>> association;	 
 };
@@ -135,13 +135,13 @@ struct Metadata {
 */
 struct SparseComponents
 {
-	SparseComponents(vector<int> rows_, vector<int> cols_, vector<double> vals_): rows(rows_), cols(cols_), vals(vals_) 
+	SparseComponents(vector<int> rows_, vector<int> cols_, vector<float> vals_): rows(rows_), cols(cols_), vals(vals_) 
 	{
 	}
 
 	vector<int> cols;
 	vector<int> rows;
-	vector<double> vals;
+	vector<float> vals;
 };
 
 
@@ -157,38 +157,38 @@ public:
 	* Constructs HierarchicalUMAP
 	*
 	* @param similarity_method_ string representing the similarity method (we only support 'euclidean' right now)
-	* @param percents_ py::array_t<double> representing the percentage of points in each hierarchy level after the first level (whole dataset)
+	* @param percents_ py::array_t<float> representing the percentage of points in each hierarchy level after the first level (whole dataset)
 	* @param n_neighbors_ int representing the number of neighbors for knn computation
-	* @param min_dist_ double representing the minimum distance between manifold structures
+	* @param min_dist_ float representing the minimum distance between manifold structures
 	* @param knn_algorithm_ string representing which knn algorithm to use
 	* @param init_ string representing the initialization of low-dimensional representation
 	* @param verbose_ bool controling the verbosity of HUMAP	
 	*/
-	HierarchicalUMAP(string similarity_method_, py::array_t<double> percents_, int n_neighbors_=15, double min_dist_=0.15, 
+	HierarchicalUMAP(string similarity_method_, py::array_t<float> percents_, int n_neighbors_=15, float min_dist_=0.15, 
 					 string knn_algorithm_="NNDescent", string init_="Spectral", bool verbose_=false, bool reproducible_=false) 
 	: similarity_method(similarity_method_), n_neighbors(n_neighbors_), min_dist(min_dist_), 
 		knn_algorithm(knn_algorithm_), percent_glue(0.0), init(init_), verbose(verbose_), reproducible(reproducible_) {
 
-		percents = vector<double>((double*)percents_.request().ptr, (double*)percents_.request().ptr + percents_.request().shape[0]);
+		percents = vector<float>((float*)percents_.request().ptr, (float*)percents_.request().ptr + percents_.request().shape[0]);
 	}
 
 	HierarchicalUMAP() {}
 
 
 	// fits the hierarchy on X
-	void fit(py::array_t<double> X, py::array_t<int> y);
+	void fit(py::array_t<float> X, py::array_t<int> y);
 
 	// returns the hierarchy level labels 
 	py::array_t<int> get_labels(int level);
 
 	// returns the subset X associated to the hierarchy level
-	Eigen::SparseMatrix<double, Eigen::RowMajor> get_data(int level);
+	Eigen::SparseMatrix<float, Eigen::RowMajor> get_data(int level);
 
 	// returns the embedding of the hierarchy level
-	py::array_t<double> get_embedding(int level);
+	py::array_t<float> get_embedding(int level);
 
 	// generates and returns the embedding of the hierarchy level
-	py::array_t<double> transform(int level);
+	py::array_t<float> transform(int level);
 
 	// returns the indices of the embedding corresponding to the hierarchy level below
 	py::array_t<int> get_indices(int level);
@@ -200,10 +200,10 @@ public:
 	py::array_t<int> get_original_indices(int level);
 
 	// generates the embedding for the hierarchy level below based on a set of classes
-	py::array_t<double> project(int level, py::array_t<int> c);	
+	py::array_t<float> project(int level, py::array_t<int> c);	
 
 	// generates the embeddding for the hierarchy level below based on a set of indices
-	py::array_t<double> project_indices(int level, py::array_t<int> indices);
+	py::array_t<float> project_indices(int level, py::array_t<int> indices);
 
 	// get the labels of the embedded subset
 	py::array_t<int> get_labels_selected() { return py::cast(this->labels_selected); }
@@ -218,7 +218,7 @@ public:
 
 	py::array_t<int> get_knn(int i) { return py::cast(this->reducers[i].knn_indices()); } 
 
-	py::array_t<double> get_knn_dists(int i) { return py::cast(this->reducers[i].knn_dists()); }
+	py::array_t<float> get_knn_dists(int i) { return py::cast(this->reducers[i].knn_dists()); }
 
 	// sets the number of random walks for landmark selection
 	void set_landmarks_nwalks(int value) { this->landmarks_nwalks = value; }
@@ -239,16 +239,16 @@ public:
 	void set_distance_similarity(bool value) { this->distance_similarity = value; }
 
 	// sets the ab parameters computed using Python
-	void set_ab_parameters(double a, double b) { this->a = a; this->b = b; }
+	void set_ab_parameters(float a, float b) { this->a = a; this->b = b; }
 
 	// defines how the embedding will be performed
 	void set_focus_context(bool value) { this->focus_context = value; }
 
 	// set how free is the fixed data points across level
-	void set_fixing_term(double fixing_term) { this->_fixing_term = fixing_term; }
+	void set_fixing_term(float fixing_term) { this->_fixing_term = fixing_term; }
 
 	// fix datapoints
-	void set_fixed_datapoints(py::array_t<double> fixed) { this->fixed_datapoints = convert_to_vector(fixed); }
+	void set_fixed_datapoints(py::array_t<float> fixed) { this->fixed_datapoints = convert_to_vector(fixed); }
 
 	// file
 	void set_info_file(string filename) { 
@@ -282,10 +282,10 @@ private:
 	bool reproducible;
 	bool low_memory = true;
 	
-	double min_dist = 0.15;
-	double a = -1.0, b = -1.0;
-	double percent_glue = 0.0;
-	double _fixing_term = 0.01;
+	float min_dist = 0.15;
+	float a = -1.0, b = -1.0;
+	float percent_glue = 0.0;
+	float _fixing_term = 0.01;
 
 	string output_filename = "";
 	
@@ -299,14 +299,14 @@ private:
 	vector<int>                    indices_selected;
 	vector<int>                    indices_fixed;
 	vector<bool>                   free_datapoints;
-	vector<double> 				   percents;
+	vector<float> 				   percents;
 	vector<vector<int>>            hierarchy_y;
 	vector<vector<int>>            original_indices;
 	// vector<vector<int>>            _indices;
-	// vector<vector<double>> 		   _sigmas;
-	vector<vector<double>> 		   fixed_datapoints;
+	// vector<vector<float>> 		   _sigmas;
+	vector<vector<float>> 		   fixed_datapoints;
 	vector<vector<int>>            level_landmarks;
-	vector<vector<vector<double>>> embeddings;
+	vector<vector<vector<float>>> embeddings;
 
 	vector<Metadata> metadata;
 
@@ -322,57 +322,57 @@ private:
 	void add_similarity(int index, int i, vector<vector<int>>& neighborhood, std::vector<std::vector<int> >& indices, 
 						int* mapper,
 						// map<int, int>& mapper, 
-						// double* elements, 
-						unordered_map<string, double>& elements,
-						// vector<vector<double>>& elements,
-						vector<vector<int>>& indices_nzeros, int n, double max_incidence, 
+						// float* elements, 
+						unordered_map<string, float>& elements,
+						// vector<vector<float>>& elements,
+						vector<vector<int>>& indices_nzeros, int n, float max_incidence, 
 						// vector<vector<int>>& association
 						vector<unordered_map<int, int>>& association);
 
 	// create a sparse represention after similarity computaiton 
 	SparseComponents create_sparse(int n, int n_neighbors, 
-			// double* elements, 
-			unordered_map<string, double>& elements,
-			// vector<vector<double>>& elements,
+			// float* elements, 
+			unordered_map<string, float>& elements,
+			// vector<vector<float>>& elements,
 			vector<vector<int>>& indices_nzeros);
 
 	// compute the similarity among landmarks
 	SparseComponents sparse_similarity(int level, int n, int n_neighbors, vector<int>& greatest, vector<vector<int>>& neighborhood,
-									   double max_incidence, vector<unordered_map<int, int>>& association
+									   float max_incidence, vector<unordered_map<int, int>>& association
 									//    vector<vector<int>>& association
 									   );
 
 	// update the position of a landmark based on its surroundings	
-	vector<double> update_position(int i, vector<int>& neighbors, umap::Matrix& X);
+	vector<float> update_position(int i, vector<int>& neighbors, umap::Matrix& X);
 
 	// performs the embedding on the dataset X using the graph force 
-	vector<vector<double>> embed_data(int level, Eigen::SparseMatrix<double, Eigen::RowMajor>& graph, umap::Matrix& X);
+	vector<vector<float>> embed_data(int level, Eigen::SparseMatrix<float, Eigen::RowMajor>& graph, umap::Matrix& X);
 
 	// associates points to landmarks
 	void associate_to_landmarks(int n, int n_neighbors, int* indices, vector<int>& cols, 
-								vector<double>& strength, vector<int>& owners, vector<int>& indices_landmark, 
+								vector<float>& strength, vector<int>& owners, vector<int>& indices_landmark, 
 								vector<vector<int>>& association, vector<int>& count_influence, vector<int>& is_landmark, 
-								vector<vector<double>>& knn_dists);
+								vector<vector<float>>& knn_dists);
 
 	// associates points to landmarks
 	void associate_to_landmarks(int n, int n_neighbors, vector<int>& landmarks, vector<int>& cols, 
-								vector<double>& strength, vector<int>& owners, vector<int>& indices, 
+								vector<float>& strength, vector<int>& owners, vector<int>& indices, 
 								vector<vector<int>>& association, vector<int>& count_influence, 
-								vector<int>& is_landmark, vector<vector<double>>& knn_dists );
+								vector<int>& is_landmark, vector<vector<float>>& knn_dists );
 
 	// searches for a point owner using the knn structure computed in UMAP
 	int depth_first_search(int n_neighbors, int* neighbors, vector<int>& cols, 
-	  					   vector<double>& strength, vector<int>& owners, vector<int>& is_landmark);
+	  					   vector<float>& strength, vector<int>& owners, vector<int>& is_landmark);
 
 	// returns the influence of each index
 	vector<int> get_influence_by_indices(int level, vector<int> indices);
 
 	// helper function to project indices
-	py::array_t<double> project_data(int level, vector<int> selected_indices);
+	py::array_t<float> project_data(int level, vector<int> selected_indices);
 
-	SparseComponents compute_landmark_similarity(vector<vector<int>>& neighborhood, double M, int n_neighbors, int N);
+	SparseComponents compute_landmark_similarity(vector<vector<int>>& neighborhood, float M, int n_neighbors, int N);
 
-	double dRNH(unordered_map<int, int>& l_u, unordered_map<int, int>& l_v);
+	float dRNH(unordered_map<int, int>& l_u, unordered_map<int, int>& l_v);
 
 	void save(string filename);
